Move DynamicListView to its own file
